+++ concurrency/v0/CheckWebsites.go
@@ -1,10 +1,12 @@
 package concurrency
 
-func CheckWebsites(urls []string) map[string]bool {
+type WebsiteChecker func(string) bool
+
+func CheckWebsites(wc WebsiteChecker, urls []string) map[string]bool {
 	results := make(map[string]bool)
 
 	for _, url := range urls {
-		results[url] = CheckWebsite(url)
+		results[url] = wc(url)
 	}
 
 	return results
