+++ concurrency/v1/CheckWebsites.go
@@ -1,11 +1,21 @@
 package concurrency
 
-func CheckWebsites(urls []string) map[string]bool {
+import (
+	"time"
+)
+
+type WebsiteChecker func(string) bool
+
+func CheckWebsites(wc WebsiteChecker, urls []string) map[string]bool {
 	results := make(map[string]bool)
 
 	for _, url := range urls {
-		results[url] = CheckWebsite(url)
+		go func(u string) {
+			results[u] = wc(u)
+		}(url)
 	}
+
+	time.Sleep(2 * time.Second)
 
 	return results
 }
