--- concurrency/v2/CheckWebsite.go
@@ -1,17 +1,0 @@
-package concurrency
-
-import "net/http"
-
-// CheckWebsite returns true if the URL returns a 200 status code, false otherwise
-func CheckWebsite(url string) bool {
-	response, err := http.Head(url)
-	if err != nil {
-		return false
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return false
-	}
-
-	return true
-}
